youbot_gazebo_publisher/src/requirements.txt
actionlib==1.11.16
adium-theme-ubuntu==0.3.4
angles==1.9.12
attrs==15.2.0
autobahn==0.10.3
base-local-planner==1.14.9
bondpy==1.8.3
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
chardet==2.3.0
controller-manager==0.13.5
controller-manager-msgs==0.13.5
controller-manager-tests==0.13.5
cryptography==1.2.3
cv-bridge==1.12.8
cycler==0.9.0
defusedxml==0.4.1
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.12
dynamic-reconfigure==1.5.50
ecdsa==0.13
empy==3.3.2
enum34==1.1.2
futures==3.0.5
gazebo-plugins==2.5.21
gazebo-ros==2.5.21
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.14
idna==2.0
image-geometry==1.12.8
interactive-markers==1.11.5
ipaddress==1.0.16
joint-state-publisher==1.12.15
laser-geometry==1.6.7
lz4==0.7.0
matplotlib==1.5.1
message-filters==1.12.17
mpi4py==1.3.1
msgpack-python==0.4.6
netifaces==0.10.4
nose==1.3.7
numpy==1.11.0
PAM==0.4.2
paramiko==1.16.0
Pillow==3.1.2
pluginlib==1.11.3
psutil==3.4.2
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pydot==1.0.29
Pygments==2.1
PyOpenGL==3.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyserial==3.0.1
Pyste==0.9.10
python-dateutil==2.4.2
python-qt-binding==0.3.7
python-snappy==0.5
pytz==2014.10
PyYAML==3.11
qt-dotgraph==0.3.17
qt-gui==0.3.17
qt-gui-cpp==0.3.17
qt-gui-py-common==0.3.17
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.12.17
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosdep==0.20.1
rosdep-modules==0.20.1
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.12.17
roslaunch==1.12.17
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.17
rosmake==1.14.6
rosmaster==1.12.17
rosmsg==1.12.17
rosnode==1.12.17
rosparam==1.12.17
rospkg==1.3.0
rospkg-modules==1.3.0
rospy==1.12.17
rosservice==1.12.17
rostest==1.12.17
rostopic==1.12.17
rosunit==1.14.6
roswtf==1.12.17
rqt-action==0.4.9
rqt-bag==0.5.0
rqt-bag-plugins==0.5.0
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.16
rqt-joint-trajectory-controller==0.13.6
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.12.17
sensor-msgs==1.12.7
service-identity==16.0.0
six==1.10.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.17
trollius==2.0.1
Twisted==16.0.0
txaio==1.0.0
unity-lens-photos==1.0
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xacro==1.11.3
zope.interface==4.1.3
